+++ domain/src/main/java/be/xplore/notifyme/domain/AvailableLanguages.java
@@ -1,0 +1,6 @@
+package be.xplore.notifyme.domain;
+
+public enum AvailableLanguages {
+  EN,
+  NL
+}
