+++ domain/src/main/java/be/xplore/notifyme/domain/SystemMessages.java
@@ -1,4 +1,4 @@
-package be.xplore.notifyme.services.systemmessages;
+package be.xplore.notifyme.domain;
 
 public enum SystemMessages {
   CANCEL_EVENT,
