+++ domain/src/main/java/be/xplore/notifyme/services/communicationstrategies/implementations/EmailCommunicationStrategy.java
@@ -1,6 +1,8 @@
-package be.xplore.notifyme.services.communicationstrategies;
+package be.xplore.notifyme.services.communicationstrategies.implementations;
 
 import be.xplore.notifyme.domain.Notification;
+import be.xplore.notifyme.services.communicationstrategies.ICommunicationStrategy;
+import be.xplore.notifyme.services.communicationstrategies.IEmailService;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -11,21 +13,23 @@ import org.keycloak.representations.account.UserRepresentation;
 @Getter
 @Setter
 @JsonSerialize
-public class SmsCommunicationStrategy implements ICommunicationStrategy {
-  private final ISmsService smsService;
+public class EmailCommunicationStrategy implements
+    ICommunicationStrategy {
+
+  private final IEmailService emailService;
 
   @Override
   public void send(Notification notification) {
-    smsService.send(notification.getCommunicationAddress(), notification.getMessage());
+    emailService.send(notification.getCommunicationAddress(), notification.getMessage());
   }
 
   @Override
   public String getCommunicationAddress(UserRepresentation userRepresentation) {
-    return userRepresentation.getAttributes().get("phone_number").get(0);
+    return userRepresentation.getEmail();
   }
 
   @Override
   public String getName() {
-    return "SMS";
+    return "Email";
   }
 }
