+++ domain/src/main/java/be/xplore/notifyme/services/communicationstrategies/implementations/SmsCommunicationStrategy.java
@@ -1,6 +1,8 @@
-package be.xplore.notifyme.services.communicationstrategies;
+package be.xplore.notifyme.services.communicationstrategies.implementations;
 
 import be.xplore.notifyme.domain.Notification;
+import be.xplore.notifyme.services.communicationstrategies.ICommunicationStrategy;
+import be.xplore.notifyme.services.communicationstrategies.ISmsService;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -11,23 +13,21 @@ import org.keycloak.representations.account.UserRepresentation;
 @Getter
 @Setter
 @JsonSerialize
-public class EmailCommunicationStrategy implements
-    ICommunicationStrategy {
-
-  private final IEmailService emailService;
+public class SmsCommunicationStrategy implements ICommunicationStrategy {
+  private final ISmsService smsService;
 
   @Override
   public void send(Notification notification) {
-    emailService.send(notification.getCommunicationAddress(), notification.getMessage());
+    smsService.send(notification.getCommunicationAddress(), notification.getMessage());
   }
 
   @Override
   public String getCommunicationAddress(UserRepresentation userRepresentation) {
-    return userRepresentation.getEmail();
+    return userRepresentation.getAttributes().get("phone_number").get(0);
   }
 
   @Override
   public String getName() {
-    return "Email";
+    return "SMS";
   }
 }
