+++ domain/src/main/java/be/xplore/notifyme/services/implementations/NotificationService.java
@@ -1,13 +1,16 @@
-package be.xplore.notifyme.services;
+package be.xplore.notifyme.services.implementations;
 
 import be.xplore.notifyme.domain.Event;
 import be.xplore.notifyme.domain.Message;
 import be.xplore.notifyme.domain.Notification;
+import be.xplore.notifyme.domain.SystemMessages;
 import be.xplore.notifyme.domain.User;
 import be.xplore.notifyme.persistence.IMessageRepo;
 import be.xplore.notifyme.persistence.INotificationRepo;
-import be.xplore.notifyme.services.systemmessages.PickLanguageService;
-import be.xplore.notifyme.services.systemmessages.SystemMessages;
+import be.xplore.notifyme.services.INotificationService;
+import be.xplore.notifyme.services.IOrganisationService;
+import be.xplore.notifyme.services.IUserService;
+import be.xplore.notifyme.services.systemmessages.implementations.PickLanguageService;
 import java.util.Collection;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -74,7 +77,7 @@ public class NotificationService implements INotificationService {
    * @param attribute   any attributes needed to create the message.
    */
   public void createAndSendSystemNotification(String userId, SystemMessages messageName,
-      Object[] attribute) {
+                                              Object[] attribute) {
     var user = userService.getUser(userId);
     var message = pickLanguageService.getLanguageService(user.getPreferedLanguage())
         .getSystemMessage(messageName, attribute);
@@ -86,12 +89,12 @@ public class NotificationService implements INotificationService {
   @Override
   public void notifyOrganisationManagersForCancelEvent(Event event, SystemMessages messageName) {
     List<User> orgManagers = organisationService.getOrganisationManagersForEvent(event.getId());
-    notifyUsers(orgManagers, messageName, new Object[]{event});
+    notifyUsers(orgManagers, messageName, new Object[] {event});
   }
 
   @Override
   public void notifyOrganisationManagersForUserEvent(Event event, User user,
-      SystemMessages messageName) {
+                                                     SystemMessages messageName) {
     List<User> orgManagers = organisationService.getOrganisationManagersForEvent(event.getId());
     orgManagers
         .removeIf(orgmgr -> orgmgr.getOrganisations().stream().noneMatch(organisationUser ->
@@ -99,7 +102,7 @@ public class NotificationService implements INotificationService {
                 cancellingOu -> cancellingOu.getOrganisationUserKey().getOrganisationId()
                     .equals(organisationUser.getOrganisationUserKey().getOrganisationId()))
                 && organisationUser.isOrganisationLeader()));
-    notifyUsers(orgManagers, messageName, new Object[]{user.getUserName(), event.getTitle()});
+    notifyUsers(orgManagers, messageName, new Object[] {user.getUserName(), event.getTitle()});
   }
 
   @Override
@@ -114,7 +117,7 @@ public class NotificationService implements INotificationService {
 
   @Override
   public void notifyUsers(Collection<User> users, SystemMessages systemMessageName,
-      Object[] attributes) {
+                          Object[] attributes) {
     if (users != null) {
       for (User user : users) {
         createAndSendSystemNotification(user.getUserId(), systemMessageName, attributes);
@@ -124,7 +127,7 @@ public class NotificationService implements INotificationService {
 
   @Override
   public void notifyOrganisationManagers(long organisationId, String sender, String title,
-      String text) {
+                                         String text) {
     var message = this.createMessage(title, text);
     var orgManagers = organisationService.getOrganisationManagers(organisationId);
     for (User orgManager : orgManagers) {
