+++ domain/src/main/java/be/xplore/notifyme/services/implementations/UserOrgApplicationService.java
@@ -1,13 +1,14 @@
-package be.xplore.notifyme.services;
+package be.xplore.notifyme.services.implementations;
 
 import be.xplore.notifyme.domain.OrgApplicationStatus;
 import be.xplore.notifyme.domain.Organisation;
 import be.xplore.notifyme.domain.OrganisationUserKey;
+import be.xplore.notifyme.domain.SystemMessages;
 import be.xplore.notifyme.domain.User;
 import be.xplore.notifyme.domain.UserOrgApplication;
 import be.xplore.notifyme.persistence.IOrganisationRepo;
+import be.xplore.notifyme.services.IUserOrgApplicationService;
 import be.xplore.notifyme.services.security.OrganisationSecurityService;
-import be.xplore.notifyme.services.systemmessages.SystemMessages;
 import java.security.Principal;
 import java.util.List;
 import lombok.AllArgsConstructor;
@@ -41,12 +42,12 @@ public class UserOrgApplicationService implements IUserOrgApplicationService {
   }
 
   private void sendUserApplicationNotificationToAllOrgManagers(Long organisationId,
-      User appliedUser) {
+                                                               User appliedUser) {
     var org = organisationService.getOrganisationIncAppliedUsers(organisationId);
     org.getUsers().forEach(ou -> notificationService
         .createAndSendSystemNotification(ou.getUser().getUserId(),
             SystemMessages.USER_APPLICATION,
-            new Object[]{ou.getUser().getUserName(), org.getName()}));
+            new Object[] {ou.getUser().getUserName(), org.getName()}));
   }
 
   @Override
@@ -78,7 +79,7 @@ public class UserOrgApplicationService implements IUserOrgApplicationService {
    */
   @Override
   public void respondToApplication(OrganisationUserKey organisationUserKey, boolean accept,
-      Principal principal) {
+                                   Principal principal) {
     if (accept) {
       var organisation = organisationService.addUserToOrganisation(organisationUserKey.getUserId(),
           organisationUserKey.getOrganisationId());
@@ -95,7 +96,7 @@ public class UserOrgApplicationService implements IUserOrgApplicationService {
   private void sendUserApplicationApprovalNotification(User user, Organisation organisation) {
     notificationService
         .createAndSendSystemNotification(user.getUserId(), SystemMessages.APPLICATION_APPROVED,
-            new Object[]{organisation.getName()});
+            new Object[] {organisation.getName()});
   }
 
   /**
@@ -106,7 +107,7 @@ public class UserOrgApplicationService implements IUserOrgApplicationService {
    * @param principal    representation of the authenticated user.
    */
   private void secureOrgManagerRequestFromPrincipal(Organisation organisation,
-      Principal principal) {
+                                                    Principal principal) {
     var user = userService.getUserFromprincipalIncOrganisations(principal);
     organisationSecurityService.checkUserIsOrgManager(user, organisation);
   }
