+++ domain/src/main/java/be/xplore/notifyme/services/implementations/UserOrgService.java
@@ -1,6 +1,7 @@
-package be.xplore.notifyme.services;
+package be.xplore.notifyme.services.implementations;
 
 import be.xplore.notifyme.domain.Organisation;
+import be.xplore.notifyme.services.IUserOrgService;
 import be.xplore.notifyme.services.security.OrganisationSecurityService;
 import java.security.Principal;
 import java.util.ArrayList;
