--- domain/src/main/java/be/xplore/notifyme/services/systemmessages/AvailableLanguages.java
@@ -1,6 +1,0 @@
-package be.xplore.notifyme.services.systemmessages;
-
-public enum AvailableLanguages {
-  EN,
-  NL
-}
