+++ domain/src/test/java/be/xplore/notifyme/services/CommunicationPreferenceServiceTest.java
@@ -17,9 +17,11 @@ import static org.mockito.Mockito.when;
 import be.xplore.notifyme.domain.CommunicationPreference;
 import be.xplore.notifyme.domain.User;
 import be.xplore.notifyme.persistence.ICommunicationPreferenceRepo;
-import be.xplore.notifyme.services.communicationstrategies.EmailCommunicationStrategy;
 import be.xplore.notifyme.services.communicationstrategies.ICommunicationStrategy;
-import be.xplore.notifyme.services.communicationstrategies.SmsCommunicationStrategy;
+import be.xplore.notifyme.services.communicationstrategies.implementations.EmailCommunicationStrategy;
+import be.xplore.notifyme.services.communicationstrategies.implementations.SmsCommunicationStrategy;
+import be.xplore.notifyme.services.implementations.CommunicationPreferenceService;
+import be.xplore.notifyme.services.implementations.KeycloakCommunicationService;
 import java.util.ArrayList;
 import javax.validation.ValidationException;
 import org.junit.jupiter.api.Test;
