+++ domain/src/test/java/be/xplore/notifyme/services/EventServiceTest.java
@@ -26,6 +26,11 @@ import be.xplore.notifyme.exception.UnauthorizedException;
 import be.xplore.notifyme.persistence.IEventRepo;
 import be.xplore.notifyme.persistence.IMessageRepo;
 import be.xplore.notifyme.persistence.IVenueRepo;
+import be.xplore.notifyme.services.implementations.EventService;
+import be.xplore.notifyme.services.implementations.NotificationService;
+import be.xplore.notifyme.services.implementations.TokenService;
+import be.xplore.notifyme.services.implementations.UserService;
+import be.xplore.notifyme.services.implementations.VenueService;
 import java.security.Principal;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
