+++ domain/src/test/java/be/xplore/notifyme/services/LineServiceTest.java
@@ -17,6 +17,10 @@ import be.xplore.notifyme.domain.Venue;
 import be.xplore.notifyme.dto.CreateLineDto;
 import be.xplore.notifyme.exception.CrudException;
 import be.xplore.notifyme.persistence.ILineRepo;
+import be.xplore.notifyme.services.implementations.EventService;
+import be.xplore.notifyme.services.implementations.FacilityService;
+import be.xplore.notifyme.services.implementations.LineService;
+import be.xplore.notifyme.services.implementations.UserService;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.HashSet;
