+++ domain/src/test/java/be/xplore/notifyme/services/TokenServiceTest.java
@@ -6,6 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import be.xplore.notifyme.exception.TokenHandlerException;
+import be.xplore.notifyme.services.implementations.TokenService;
 import com.c4_soft.springaddons.security.oauth2.test.annotations.OidcStandardClaims;
 import com.c4_soft.springaddons.security.oauth2.test.annotations.keycloak.WithMockKeycloakAuth;
 import org.junit.jupiter.api.Test;
