+++ domain/src/test/java/be/xplore/notifyme/services/UserOrgApplicationServiceTest.java
@@ -19,6 +19,10 @@ import be.xplore.notifyme.domain.User;
 import be.xplore.notifyme.domain.UserOrgApplication;
 import be.xplore.notifyme.exception.OrgApplicationNotFoundException;
 import be.xplore.notifyme.persistence.IOrganisationRepo;
+import be.xplore.notifyme.services.implementations.NotificationService;
+import be.xplore.notifyme.services.implementations.OrganisationService;
+import be.xplore.notifyme.services.implementations.UserOrgApplicationService;
+import be.xplore.notifyme.services.implementations.UserService;
 import be.xplore.notifyme.services.security.OrganisationSecurityService;
 import java.util.ArrayList;
 import java.util.LinkedList;
