+++ domain/src/test/java/be/xplore/notifyme/services/UserOrgServiceTest.java
@@ -12,6 +12,9 @@ import be.xplore.notifyme.domain.Organisation;
 import be.xplore.notifyme.domain.OrganisationUser;
 import be.xplore.notifyme.domain.User;
 import be.xplore.notifyme.exception.UnauthorizedException;
+import be.xplore.notifyme.services.implementations.OrganisationService;
+import be.xplore.notifyme.services.implementations.UserOrgService;
+import be.xplore.notifyme.services.implementations.UserService;
 import be.xplore.notifyme.services.security.OrganisationSecurityService;
 import java.util.List;
 import java.util.Optional;
