+++ domain/src/test/java/be/xplore/notifyme/services/VenueServiceTest.java
@@ -21,6 +21,9 @@ import be.xplore.notifyme.exception.CrudException;
 import be.xplore.notifyme.exception.SaveToDatabaseException;
 import be.xplore.notifyme.exception.TokenHandlerException;
 import be.xplore.notifyme.persistence.IVenueRepo;
+import be.xplore.notifyme.services.implementations.TokenService;
+import be.xplore.notifyme.services.implementations.UserService;
+import be.xplore.notifyme.services.implementations.VenueService;
 import java.security.Principal;
 import java.util.ArrayList;
 import java.util.HashSet;
