+++ domain/src/test/java/be/xplore/notifyme/services/communicationstrategies/EmailCommunicationStrategyTest.java
@@ -5,6 +5,7 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.instanceOf;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
+import be.xplore.notifyme.services.communicationstrategies.implementations.EmailCommunicationStrategy;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
