+++ domain/src/test/java/be/xplore/notifyme/services/systemmessages/PickLanguageServiceTest.java
@@ -3,6 +3,10 @@ package be.xplore.notifyme.services.systemmessages;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.instanceOf;
 
+import be.xplore.notifyme.domain.AvailableLanguages;
+import be.xplore.notifyme.services.systemmessages.implementations.PickLanguageService;
+import be.xplore.notifyme.services.systemmessages.implementations.SystemMessagesEn;
+import be.xplore.notifyme.services.systemmessages.implementations.SystemMessagesNl;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
