+++ rest-api/src/test/java/be/xplore/notifyme/controller/OrganisationManagerControllerTest.java
@@ -18,9 +18,9 @@ import be.xplore.notifyme.domain.UserOrgApplication;
 import be.xplore.notifyme.exception.GeneralExceptionHandler;
 import be.xplore.notifyme.exception.UnauthorizedException;
 import be.xplore.notifyme.services.ITeamApplicationService;
-import be.xplore.notifyme.services.UserOrgApplicationService;
-import be.xplore.notifyme.services.UserOrgService;
-import be.xplore.notifyme.services.UserService;
+import be.xplore.notifyme.services.implementations.UserOrgApplicationService;
+import be.xplore.notifyme.services.implementations.UserOrgService;
+import be.xplore.notifyme.services.implementations.UserService;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
