+++ rest-api/src/test/java/be/xplore/notifyme/controller/VenueManagerControllerTest.java
@@ -21,11 +21,11 @@ import be.xplore.notifyme.domain.Venue;
 import be.xplore.notifyme.dto.CreateEventDto;
 import be.xplore.notifyme.exception.CrudException;
 import be.xplore.notifyme.exception.GeneralExceptionHandler;
-import be.xplore.notifyme.services.EventService;
-import be.xplore.notifyme.services.FacilityService;
 import be.xplore.notifyme.services.IUserService;
-import be.xplore.notifyme.services.LineService;
-import be.xplore.notifyme.services.VenueService;
+import be.xplore.notifyme.services.implementations.EventService;
+import be.xplore.notifyme.services.implementations.FacilityService;
+import be.xplore.notifyme.services.implementations.LineService;
+import be.xplore.notifyme.services.implementations.VenueService;
 import java.security.Principal;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
